refactor: Move Request and RequestBuilder structs to common_utils crate
