+++ crates/common_utils/Cargo.toml
@@ -19,17 +19,20 @@ diesel = "2.1.0"
 error-stack = "0.3.1"
 futures = { version = "0.3.28", optional = true }
 hex = "0.4.3"
+http = "0.2.9"
 md5 = "0.7.0"
 nanoid = "0.4.0"
 once_cell = "1.18.0"
 quick-xml = { version = "0.28.2", features = ["serialize"] }
 rand = "0.8.5"
 regex = "1.8.4"
+reqwest = { version = "0.11.18", features = ["json", "native-tls", "gzip", "multipart"] }
 ring = { version = "0.16.20", features = ["std"] }
 serde = { version = "1.0.163", features = ["derive"] }
 serde_json = "1.0.96"
 serde_urlencoded = "0.7.1"
 signal-hook = { version = "0.3.15", optional = true }
+strum = { version = "0.24.1", features = ["derive"] }
 thiserror = "1.0.40"
 time = { version = "0.3.21", features = ["serde", "serde-well-known", "std"] }
 tokio = { version = "1.28.2", features = ["macros", "rt-multi-thread"], optional = true }
